+++ packages/core/docs/commercetools/changelog.md
@@ -2,6 +2,8 @@
 
 ## 0.2.6 - not released
 
+- The address `contactInfo` field is deprecated in the CT api. We have added support for the contact information fields directly in the address and will now show a warning when deprecated field is used ([#5083](https://github.com/DivanteLtd/vue-storefront/pull/5083))
+
 ## 0.2.5
 
 - `customQuery` for checkout composables ([#5025](https://github.com/DivanteLtd/vue-storefront/issues/5025))